+++ demo-app/src/main/java/io/opentelemetry/android/demo/ui/cart/CartFragment.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.demo.ui.home
+package io.opentelemetry.android.demo.ui.cart
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -12,18 +12,18 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import io.opentelemetry.android.demo.databinding.FragmentHomeBinding
+import io.opentelemetry.android.demo.databinding.FragmentNotificationsBinding
 
-class HomeFragment : Fragment() {
+class CartFragment : Fragment() {
     // Renamed from _binding (default) due to ktlint problem below
-    private var binding: FragmentHomeBinding? = null
+    private var binding: FragmentNotificationsBinding? = null
 
     // This property is only valid between onCreateView and onDestroyView.
     // Currently failing ktlint due to
     // https://github.com/pinterest/ktlint/issues/2448
     // which hasn't hit the aging gradle spotless plugin yet
 //    private val binding get() = _binding!!
-    private fun getBinding(): FragmentHomeBinding {
+    private fun getBinding(): FragmentNotificationsBinding {
         return binding!!
     }
 
@@ -32,14 +32,14 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        val homeViewModel =
-            ViewModelProvider(this).get(HomeViewModel::class.java)
+        val cartViewModel =
+            ViewModelProvider(this).get(CartViewModel::class.java)
 
-        binding = FragmentHomeBinding.inflate(inflater, container, false)
+        binding = FragmentNotificationsBinding.inflate(inflater, container, false)
         val root: View = getBinding().root
 
-        val textView: TextView = getBinding().textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = getBinding().textNotifications
+        cartViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root